__init__ tidy up & ModelSpecialSlicers to BaseModel
