+++ hyperspy/models/model1D.py
@@ -21,7 +21,7 @@ import warnings
 import numpy as np
 from traits.trait_errors import TraitError
 
-from hyperspy.model import BaseModel, ModelComponents
+from hyperspy.model import BaseModel, ModelComponents, ModelSpecialSlicers
 import hyperspy.drawing.spectrum
 from hyperspy.drawing.utils import on_figure_window_close
 from hyperspy._signals.eels import Spectrum
@@ -32,9 +32,6 @@ from hyperspy.axes import AxesManager
 from hyperspy.drawing.widgets import (DraggableVerticalLine,
                                       DraggableLabel)
 from hyperspy.gui.tools import ComponentFit
-from hyperspy import components
-from hyperspy.misc.export_dictionary import parse_flag_string
-from hyperspy.misc.slicing import copy_slice_from_whitelist
 
 
 class Model1D(BaseModel):
@@ -172,21 +169,6 @@ class Model1D(BaseModel):
         self._suspend_update = False
         self._model_line = None
         self._adjust_position_all = None
-        self._plot_components = False
-        self._whitelist = {
-            'channel_switches': None,
-            'convolved': None,
-            'free_parameters_boundaries': None,
-            'low_loss': ('sig', None),
-            'chisq.data': None,
-            'dof.data': None
-        }
-        self._slicing_whitelist = {
-            'channel_switches': 'isig',
-            'low_loss': 'inav',
-            'chisq.data': 'inav',
-            'dof.data': 'inav'}
-
         self.axis = self.axes_manager.signal_axes[0]
         self.axes_manager.connect(self.fetch_stored_values)
         self.channel_switches = np.array([True] * len(self.axis.axis))
@@ -207,6 +189,18 @@ class Model1D(BaseModel):
             self._load_dictionary(dictionary)
         self.inav = ModelSpecialSlicers(self, True)
         self.isig = ModelSpecialSlicers(self, False)
+        self._whitelist = {
+            'channel_switches': None,
+            'convolved': None,
+            'free_parameters_boundaries': None,
+            'low_loss': ('sig', None),
+            'chisq.data': None,
+            'dof.data': None}
+        self._slicing_whitelist = {
+            'channel_switches': 'isig',
+            'low_loss': 'inav',
+            'chisq.data': 'inav',
+            'dof.data': 'inav'}
 
     @property
     def spectrum(self):
@@ -850,75 +844,3 @@ class Model1D(BaseModel):
             cf.edit_traits()
         else:
             cf.apply()
-
-
-class ModelSpecialSlicers(object):
-
-    def __init__(self, model, isNavigation):
-        self.isNavigation = isNavigation
-        self.model = model
-
-    def __getitem__(self, slices):
-        array_slices = self.model.spectrum._get_array_slices(
-            slices,
-            self.isNavigation)
-        _spectrum = self.model.spectrum._slicer(slices, self.isNavigation)
-        if _spectrum.metadata.Signal.signal_type == 'EELS':
-            _model = _spectrum.create_model(
-                auto_background=False,
-                auto_add_edges=False)
-        else:
-            _model = _spectrum.create_model()
-
-        dims = self.model.axes_manager.navigation_dimension, self.model.axes_manager.signal_dimension
-        if self.isNavigation:
-            _model.channel_switches[:] = self.model.channel_switches
-        else:
-            _model.channel_switches[:] = \
-                np.atleast_1d(
-                    self.model.channel_switches[tuple(array_slices[-dims[1]:])])
-
-        twin_dict = {}
-        for comp in self.model:
-            init_args = {}
-            for k, v in comp._whitelist.iteritems():
-                if v is None:
-                    continue
-                flags_str, value = v
-                if 'init' in parse_flag_string(flags_str):
-                    init_args[k] = value
-            _model.append(getattr(components, comp._id_name)(**init_args))
-        copy_slice_from_whitelist(self.model,
-                                  _model,
-                                  dims,
-                                  (slices, array_slices),
-                                  self.isNavigation,
-                                  )
-        for co, cn in zip(self.model, _model):
-            copy_slice_from_whitelist(co,
-                                      cn,
-                                      dims,
-                                      (slices, array_slices),
-                                      self.isNavigation)
-            for po, pn in zip(co.parameters, cn.parameters):
-                copy_slice_from_whitelist(po,
-                                          pn,
-                                          dims,
-                                          (slices, array_slices),
-                                          self.isNavigation)
-                twin_dict[id(po)] = ([id(i) for i in list(po._twins)], pn)
-
-        for k in twin_dict.keys():
-            for tw_id in twin_dict[k][0]:
-                twin_dict[tw_id][1].twin = twin_dict[k][1]
-
-        _model.chisq.data = _model.chisq.data.copy()
-        _model.dof.data = _model.dof.data.copy()
-        _model.fetch_stored_values()  # to update and have correct values
-        if not self.isNavigation:
-            for _ in _model.axes_manager:
-                _model._calculate_chisq()
-
-        return _model
-
-# vim: textwidth=80
